Overhaul of alert template
The two ideas are that it should almost work out of the box and that it should
be really easy to drop something cool in. I also wanted to demonstrate both
app-level attributes and alert-level attributes, so I put the title at the app
level and the message at the action level.  In the action I built a template
to safe_substitute in any alert properties, so a message can be written with
variable like $source and they will get substituted in the message.

For demo purposes, replacing the two writes to stderr with something
substantive will allow a real action to take place.  I've been using this to
pop up an alert on OSX.  Seems to work well.

applescript = 'display notification "{}" with title "{}"'.format(message_text, title);
call(["osascript", "-e", applescript])